TileMap Initialisiation Done
I tried to stick to your format and practices, sorry if I broke stuff :(

Created AfterContentLoad Method in TileMap for Initialising Array,
cannot do this in the constructor as it creates a circular dependency
due to textures being required for initial tile drawing. NOTE: Requires
the Texture storage system to be set up with the correct texture (Grass
or blank) to be fed to the initialising tile.

Created an instance of the tilemap class (world) in GameClass, called
it's AfterContentLoad() function at the end of the ContentLoad()
function.

Added the Tile_Width constant to GameClass, just gonna use one set of
tiles until everything works, then we can branch and break!

Added Draw method to tileMap.

Hope I didn't break too much :S