IT WORKS.
Oh my glob. Just run it.

Probably most importantly, I broke some stuff to make this work:

In the menu state, it instantly goes to the ingame state in the onEnter
method, just so we can get game stuff working without worrying about
menus!

The Level class is created in the InGame() State, whish is no where near
where we need it. We need to create them all at some point, probably in
the loadcontent() method, and then we need to pass them through to the
menu state. At that point, the selected one will be passed through to
the InGameState as it should be, to be used in the tilemap generation.

Updated Tilemap class so now it interfaces with everything perfectly, as
you can see :D Do we need the Width & Height perameters now? If so we'll
need to feed them from the GameGlass.

K, it's 2:46am and I've spent the last 2 hours fixing bugs with the
loadLevel class. It should all work now. It fucking better ;_;