Continuing to add a pathfinding/moving system.
Added a lot of stuff. We now use two methods to move. Basically what
happens is one method finds the next tile to move to and the other moves
smoothly (not yet) to it. When the position of the unit matches the
position of the tile origin, the cycle is repeated until the
source/mouse click/final destination is reached. Honestly, I have no
idea whether this will work so I'm gonna branch