Continued work on movement.
Now actually moves, but is very crashy. It actually moves too fast for
us to see as you can see in the console that it hits the right limit
(I've made it just move right as no field is being generated, so the
conditions on the other three directions fail. The other major problem
is that there is no stopping code really set up yet: it keeps hitting
the right limit slowing it down into a lagfest. this needs to be fixed.
when this has been done the only other thing is that the move code needs
to be changed (move method in unit) so that it actually moves nicely and
not jumping from tile to tile. This will be part of the mouse stuff so
no worries as of yet.