Fixed health bars
Turns out the problem was that only the moving units were being updated,
this took me a good hour to find. lol.

Also, you might notice that units now die after a little while and pop
out of existence. This was for demo purposes, to disable this there is a
comment in the Update method of health entity.