Fixed Problems with camera/pathfinding, LOOK!
Hi Joe! This is a bit important so you may want to read. Basically, the
problem with all the pathfinding/camera/input was that we were always
getting the mouse position relative to the screen, so even if you were
at the bottom right of the map, your mouse position was still 0, 0 at
the top right of the screen >_< I added an optional method within Input
to fix this, and my reasons for it being the way it is are described
within it's Function Description. Enjoy Pathfinding!!